Main program, move duplicates to a folder
This program checks if every picture in cleanup is already in archive,
in case it is a duplicate it will be moved to the temp-trash folder.
This is done with a dictionary with the md5 and file name of the content
in archive, and a loop for every picture to be organize being checked
if exists in the dictionary.